requirements.txt
# LED strip game requirements
adafruit-circuitpython-neopixel
pygame>=2.5.0
